+++ src/Web/Rep/SharedReps.hs
@@ -3,6 +3,7 @@
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE QuasiQuotes #-}
 {-# OPTIONS_GHC -Wno-name-shadowing #-}
+{-# OPTIONS_GHC -Wno-x-partial #-}
 
 -- | Various SharedRep instances for common html input elements.
 module Web.Rep.SharedReps
@@ -62,7 +63,7 @@ import Prelude as P
 
 -- | Create a sharedRep from an Input.
 repInput ::
-  (Monad m, Show a) =>
+  (Monad m) =>
   -- | Parser
   (ByteString -> Either ByteString a) ->
   -- | Printer
@@ -72,12 +73,12 @@ repInput ::
   -- | initial value
   a ->
   SharedRep m a
-repInput p pr i = register p pr (\n v -> inputToHtml $ #inputVal .~ v $ #inputId .~ n $ i)
+repInput p pr i = register p pr (\n v -> markupInput pr $ #inputVal .~ v $ #inputId .~ n $ i)
 
 -- | Like 'repInput', but does not put a value into the HashMap on instantiation, consumes the value when found in the HashMap, and substitutes a default on lookup failure
-repMessage :: (Monad m, Show a) => (ByteString -> Either ByteString a) -> (a -> ByteString) -> Input a -> a -> a -> SharedRep m a
-repMessage p _ i def a =
-  message p (\n v -> inputToHtml $ #inputVal .~ v $ #inputId .~ n $ i) a def
+repMessage :: (Monad m) => (ByteString -> Either ByteString a) -> (a -> ByteString) -> Input a -> a -> a -> SharedRep m a
+repMessage p pr i def a =
+  message p (\n v -> markupInput pr $ #inputVal .~ v $ #inputId .~ n $ i) a def
 
 -- | double slider
 --
@@ -129,7 +130,7 @@ sliderV label l u s v =
 -- sliderI (Just "label") 0 1000 10 300
 --   :: (Monad m, ToHtml a, P.Integral a, Show a) => SharedRep m a
 sliderI ::
-  (Monad m, P.Integral a, Show a) =>
+  (Monad m, P.Integral a, ToByteString a) =>
   Maybe ByteString ->
   a ->
   a ->
@@ -139,13 +140,13 @@ sliderI ::
 sliderI label l u s v =
   repInput
     (runParserEither (fromIntegral <$> int))
-    (strToUtf8 . show)
-    (Input v label mempty (Slider [Attr "min" (strToUtf8 $ show l), Attr "max" (strToUtf8 $ show u), Attr "step" (strToUtf8 $ show s)]))
+    toByteString
+    (Input v label mempty (Slider [Attr "min" (toByteString l), Attr "max" (toByteString u), Attr "step" (toByteString s)]))
     v
 
 -- | integral slider with shown value
 sliderVI ::
-  (Monad m, P.Integral a, Show a) =>
+  (Monad m, P.Integral a, ToByteString a) =>
   Maybe ByteString ->
   a ->
   a ->
@@ -155,8 +156,8 @@ sliderVI ::
 sliderVI label l u s v =
   repInput
     (runParserEither (fromIntegral <$> int))
-    (strToUtf8 . show)
-    (Input v label mempty (SliderV [Attr "min" (strToUtf8 $ show l), Attr "max" (strToUtf8 $ show u), Attr "step" (strToUtf8 $ show s)]))
+    toByteString
+    (Input v label mempty (SliderV [Attr "min" (toByteString l), Attr "max" (toByteString u), Attr "step" (toByteString s)]))
     v
 
 -- | textbox classique
@@ -200,7 +201,7 @@ colorPicker label v =
 
 -- | dropdown box
 dropdown ::
-  (Monad m, Show a) =>
+  (Monad m) =>
   -- | parse an a from ByteString
   (ByteString -> Either ByteString a) ->
   -- | print an a to ByteString
@@ -221,7 +222,7 @@ dropdown p pr label opts v =
 
 -- | dropdown box with multiple selections
 dropdownMultiple ::
-  (Monad m, Show a) =>
+  (Monad m) =>
   -- | parse an a from ByteString
   Parser ByteString a ->
   -- | print an a to ByteString
@@ -251,7 +252,7 @@ datalist label opts v id'' =
 
 -- | A dropdown box designed to help represent a haskell sum type.
 dropdownSum ::
-  (Monad m, Show a) =>
+  (Monad m) =>
   (ByteString -> Either ByteString a) ->
   (a -> ByteString) ->
   Maybe ByteString ->
@@ -347,7 +348,7 @@ checkboxShow label id' v =
     zoom _2 (modify (HashMap.insert name (bool "false" "true" v)))
     pure $
       Rep
-        (inputToHtml (Input v label name (Checkbox v)) <> scriptToggleShow name id')
+        (markupInput (strToUtf8 . show) (Input v label name (Checkbox v)) <> scriptToggleShow name id')
         ( \s ->
             ( s,
               join $
@@ -395,9 +396,9 @@ accordionBoolList title prefix bodyf checkf labels xs = SharedRep $ do
             . zipWith (\l (ch, a) -> (l, a, ch)) labels
         )
         ( foldr
-            (\a x -> bimap (:) (:) a <<*>> x)
+            ((\a x -> bimap (:) (:) a <<*>> x) . (\(ch, a) -> bimap (,) (,) (checkf ch) <<*>> bodyf a))
             (pure [])
-            ((\(ch, a) -> bimap (,) (,) (checkf ch) <<*>> bodyf a) <$> xs)
+            xs
         )
   h' <- zoom _1 h
   pure (Rep (maybe mempty (elementc "h5" []) title <> h') fa)
@@ -434,21 +435,22 @@ listRep t p brf srf n defa as =
       (defaultListLabels n)
       (take n (((True,) <$> as) <> repeat (False, defa)))
 
--- a sensible default for the accordion row labels for a list
+-- | A sensible default for the accordion row labels for a list
 defaultListLabels :: Int -> [ByteString]
 defaultListLabels n = (\x -> "[" <> strToUtf8 (show x) <> "]") <$> [0 .. n] :: [ByteString]
 
 -- | Parse from a textbox
 --
 -- Uses focusout so as not to spam the reader.
-readTextbox :: (Monad m, Read a, Show a) => Maybe ByteString -> a -> SharedRep m (Either ByteString a)
-readTextbox label v = parsed . utf8ToStr <$> textbox' label (strToUtf8 $ show v)
+readTextbox :: (Monad m, Read a, ToByteString a) => Maybe ByteString -> a -> SharedRep m (Either ByteString a)
+readTextbox label v = parsed . utf8ToStr <$> textbox' label (toByteString v)
   where
     parsed str =
       case reads str of
         [(a, "")] -> Right a
         _badRead -> Left (strToUtf8 str)
 
+-- | Dropdown representation of a multi-element list.
 repChoice :: (Monad m) => Int -> [(ByteString, SharedRep m a)] -> SharedRep m a
 repChoice initt xs =
   bimap hmap mmap dd
@@ -477,6 +479,7 @@ repItemsSelect :: (Monad m) => [ByteString] -> [ByteString] -> SharedRep m [Byte
 repItemsSelect initial full =
   dropdownMultiple (strToUtf8 <$> some (satisfy (`notElem` ([','] :: [Char])))) id (Just "items") full initial
 
+-- | subtype Html class.
 subtype :: ByteString -> ByteString -> [Attr]
 subtype origt t =
   [ Attr "class" "subtype ",