+++ src/Web/Rep/Socket.hs
@@ -43,7 +43,7 @@ module Web.Rep.Socket
 where
 
 import Box
-import Box.Socket (serverApp)
+import Box.Websocket (serverApp)
 import Control.Concurrent.Async
 import Control.Monad
 import Control.Monad.State.Lazy
@@ -55,7 +55,7 @@ import Data.HashMap.Strict as HashMap
 import Data.Profunctor
 import Data.String.Interpolate
 import Data.Text (Text)
-import Data.Text qualified as Text
+import Data.Text.Encoding
 import FlatParse.Basic
 import GHC.Generics
 import MarkupParse
@@ -70,51 +70,46 @@ import Web.Rep.Shared
 import Web.Rep.SharedReps
 import Web.Scotty (middleware, scotty)
 
-toText_ :: ByteString -> Text
-toText_ = Text.pack . utf8ToStr
-
-fromText_ :: Text -> ByteString
-fromText_ = strToUtf8 . Text.unpack
-
 -- | Page with all the trimmings for a sharedRep Box
 socketPage :: Page
 socketPage =
   mempty
     & #jsOnLoad
-      .~ mconcat
-        [ webSocket,
-          runScriptJs,
-          refreshJsbJs,
-          preventEnter
-        ]
+    .~ mconcat
+      [ webSocket,
+        runScriptJs,
+        refreshJsbJs,
+        preventEnter
+      ]
 
+-- | Bootstrapped base page for a web socket.
 defaultSocketPage :: Page
 defaultSocketPage =
   bootstrapPage
     <> socketPage
-    & set
-      #htmlBody
-      ( element
-          "div"
-          [Attr "class" "container"]
-          ( element
-              "div"
-              [Attr "class" "row"]
-              (elementc "h1" [] "web-rep testing")
-              <> element
+      & set
+        #htmlBody
+        ( element
+            "div"
+            [Attr "class" "container"]
+            ( element
                 "div"
                 [Attr "class" "row"]
-                ( mconcat $
-                    ( \(t, h) ->
-                        element
-                          "div"
-                          [Attr "class" "row"]
-                          (element "h2" [] (elementc "div" [Attr "id" t] h))
-                    )
-                      <$> sections
-                )
-          )
-      )
+                (elementc "h1" [] "web-rep testing")
+                <> element
+                  "div"
+                  [Attr "class" "row"]
+                  ( mconcat $
+                      ( \(t, h) ->
+                          element
+                            "div"
+                            [Attr "class" "row"]
+                            (element "h2" [] (elementc "div" [Attr "id" t] h))
+                      )
+                        <$> sections
+                  )
+            )
+        )
   where
     sections =
       [ ("input", mempty),
@@ -169,7 +164,7 @@ codeBoxWith cfg =
     (view #codeBoxEmitterQueue cfg)
     (view #codeBoxCommitterQueue cfg)
     ( serveSocketBox (view #codeBoxSocket cfg) (view #codeBoxPage cfg)
-        . dimap (either error id . runParserEither parserJ . fromText_) (mconcat . fmap (toText_ . code))
+        . dimap (either error id . runParserEither parserJ . encodeUtf8) (mconcat . fmap (decodeUtf8 . code))
     )
 
 -- | Turn the default configuration into a live (Codensity) CodeBox
@@ -293,6 +288,7 @@ parserJ = do
 
 -- * code messaging
 
+-- | A simple schema for code that communicates changes to a Html page via JS code.
 data Code
   = Replace ByteString ByteString
   | Append ByteString ByteString
@@ -301,6 +297,7 @@ data Code
   | Val ByteString
   deriving (Eq, Show, Generic, Read)
 
+-- | Convert 'Code' to a 'ByteString'
 code :: Code -> ByteString
 code (Replace i t) = replace i t
 code (Append i t) = append i t
@@ -308,9 +305,11 @@ code (Console t) = console t
 code (Eval t) = t
 code (Val t) = val t
 
+-- | write to the console
 console :: ByteString -> ByteString
 console t = [i| console.log(#{t}) |]
 
+-- | send arbitrary byestrings.
 val :: ByteString -> ByteString
 val t = [i| jsb.ws.send(#{t}) |]
 
@@ -334,6 +333,7 @@ append d t =
      refreshJsb();
      |]
 
+-- | Double backslash newline and single quotes.
 clean :: ByteString -> ByteString
 clean =
   C.intercalate "\\'"