+++ ChangeLog.md
@@ -1,3 +1,10 @@
+0.12
+===
+
+* markupInput replaces inputToHtml as per markup-parse
+* ToByteString introduced
+* upgrade to box-socket-0.5
+
 0.11
 ===
 * Removed clay, lucid as dependencies